Maintaining the consistency in regards to changes in db migration

+++ src/main/java/com/uber/UberReviewService/dtos/ReviewDto.java
@@ -12,7 +12,7 @@ import java.util.Date;
 @NoArgsConstructor
 @Builder
 public class ReviewDto {
-    private int reviewId;
+    private Long reviewId;
     private Date createdAt;
     private Date updatedAt;
     private Double rating;
